Add a empty funciton: elm_win_profiles_set()
This function is used in Camera application and in isf.

I think we will not need this function, if we are testing in
desktop we should set out envirement to mobile to have
the same result as in device.

So when possible we should remove all calls of this function.