PTREL-546 fix: enable wayland build
The display server based on X or Wayland, is now a choice of different profile in Tizen 3.0.
Basically and consistently, two macros were used "with wayland" and "with x".
Below summarize the combination of the macros:

    | wayland |  x  | meaning
    |---------------------------
    |   0     |  1  | pure X11 platform(no wayland)
    |   1     |  0  | pure wayland platform (no X11)
    |   1     |  1  | wayland but X compatibility
    |   0     |  0  | no X and no wayland

This method unifies the meaning and usage. Deploy this method to elementary package.

Change-Id: I761ef5a9b03493fe7e867282d4d3cd2a812d7482
Signed-off-by: Eduardo Lima (Etrunko) <eduardo.lima@intel.com>

Conflicts:
	packaging/elementary.spec