Elm_Entry: set handler style according to current widget style
Previously, entry set handlers' style as default when they are created
and changed style on theme_apply function. But in usual scenario, users create
entry and set style before handlers are created, so when they are created
the style is set as default regardless of the style users already set.
This patch fixes it.

@fix

Change-Id: Icc9b2e1ca8e14349c9201df369072539906052b8