layout: apply widget's states when elm_layout_file_set is called
when a layout's theme is changed by elm_layout_theme_set, the edje is updated
according to the widget's states such as 'disabled' or 'highlight_in_theme'.
but the update is dropped for elm_layout_file_set.
this patch tied up the codes into _visuals_refresh function
so that the same update happens for elm_layout_file_set.

@fix

Change-Id: Ied92e12a42581a6bf2e0c1866cef009ddd21108b