elm_spinner: add feature in spinner to change values while dragging relative to the speed of dragging
Summary:
Currently the values in spinner change while dragging only based on
the amount of pixels dragged, this patch will enable Spinner to change value
based on the speed of dragging so that it will be more user friendly. This
will help users to alter drag values by big amounts if dragged in a good speed.

Test Plan: test_spinner.c in elementary_test

Reviewers: raster, prince.dubey, shilpasingh, cedric

Reviewed By: cedric

Subscribers: poornima.srinivasan, rajeshps, govi

Differential Revision: https://phab.enlightenment.org/D2659

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>

Conflicts:

	AUTHORS
	src/lib/elm_authors.h

Change-Id: If7f0fe979f81f32f065b0da2bbc7cb7fc28301ac