focus: add the focus region show mode
When the focus is move to the object in scroller, it is scrolled
to show the focus region as a widget.
If the focus region want to be shown as an item,
set the mode ELM_FOCUS_REGION_SHOW_ITEM.
If then, it will be scrolled as an item.
TODO: Widgets have items are added on_focus_region function.

@feature

Conflicts:
	src/bin/test_gengrid.c
	src/lib/elm_focus.h
	src/lib/elm_gengrid.eo
	src/lib/elm_widget.eo

Change-Id: Ibc66f2d10e720d90e7da328d1a3230813c72846b
origin: upstream