wayland cnp: check selection owner before calling losscb
The selection is cleared by losscb when we set it.
As result, selection is not done.
To avoid it, we should check if the new selection owner
is current selection owner or not and only call losscb
if the new selection owner is not the current one.

Test Plan:
In wayland, select text in entry (e.g double click, ctrl-a).

upstream commit: 6c13b8db9a7dbbd417caf06d9f855986d815d075
Author: Thiep Ha <thiepha@gmail.com>
Date:   Thu Dec 17 18:13:17 2015 +0000

Change-Id: I9ff64766130858f7b98613f18c4fb88a14100cc0