elm win: fix window redraw if norender is enabled on withdrawn windows
we dont set this config by default so we dont see this - but a window
starts as withdrawn then becomes normal. this means the window starts
not rendering and may take 0.5 seconds to START because the eval is
deferred by 0.5 sec to avoid rapid switch to/from norender and other
states.

@fix

Change-Id: I7342bd9e6d489268108a0d28eaba262d82b1973e