popup: If popup's parent is resized, popup need to resized, too
When window's size is not enough to show popup, popup resize smaller to show it in appropriate size.
But if window resize larger enough to show popup, popup is locked in small size.
On the contrary to this case, if window resize too smaller, popup has still large size, so popup doesn't display properly.

Change-Id: Iea7587c63e26940444d3e82f685f68d6b604d74f
Signed-off-by: Jinyong Park <j4939.park@samsung.com>