widget: set the parent-child relation again, when the widget is unset
Summary:
When the widget is unset from any container, a parent of the widget
doesn't exist. So we should set its parent to the top object.
But if we just set sd->parent, the parent can not find the widget as a
child. So the container widgets set the parent-child relation when
sub object is unset.
This commit is related to 0822ad2195d335d65208856e6e590ac47691d920.

@fix

Change-Id: Ie274db13a6513fd8ca6de3a7558bbe75f4c1b3b3