[elm_spinner] Forcefully setting highlight frame on spinner entry as on entry activation frame goes to window again.
When entry is activated in spinner, focus is forcefully being set on entry, becuase if it is
not set then according to focus cycle focus will go on window. Same way frame also goes on
window when entry of the spinner is being activated, hence forcefully setting highlight frame
on spinner entry.

Change-Id: I62a34e46d7f1af4117c53ee742785e9d1eae0003