elm_win: loop auto throttle - only do this once a window has been shown etc
only start throttling ocne we have a window and it has rendered or
been focused or mouse has gone in - then we are surely really visible
and sane and only start throttling if no windows are visible then. if
auto throttle is on this can improve startup time by the trottle amount

By Carsten Haitzler.

Signed-off-by: Hosang Kim <hosang12.kim@samsung.com>
Change-Id: I206e6d7e7b3aa21e7dc8d4071b8a515c62fbfbeb